otel-lab/app/requirements.txt
flask
prometheus_client
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
